Fix 64-bit timestamp formatting on 32-bit targets (#23997)
fix: print 64-bit time_t with PRId64 on 32-bit targets

On armv7l/armhf/static builds long is 32-bit but time_t is 64-bit, so
passing a time_t to %ld desynchronises va_arg and a later %s can be
handed an integer as a pointer (the STREAM:18 crash class, #23397).
Crash-capable sites: the dbengine page-conflict warning, extent read
errors, invalid replay timestamps, and a journal mmap-failure log.

Use %" PRId64 " with an (int64_t) cast for every time_t, %zu for
size_t/Word_t. Format strings only; output on 64-bit is unchanged.
An i386 _TIME_BITS=64 build goes from 145 -Wformat mismatches to 0.

+++ src/daemon/unit_test.c
@@ -2044,16 +2044,16 @@ static int test_rrdset_rejects_invalid_update_every(void) {
     uint32_t min_update_every =
         __atomic_load_n(&st->rrdhost->stream.rcv.min_update_every, __ATOMIC_RELAXED);
     if(min_update_every != valid_update_every) {
-        fprintf(stderr, "%s: valid update every selected minimum %u instead of %ld seconds\n",
-                __FUNCTION__, min_update_every, valid_update_every);
+        fprintf(stderr, "%s: valid update every selected minimum %u instead of %" PRId64 " seconds\n",
+                __FUNCTION__, min_update_every, (int64_t)valid_update_every);
         rc = 1;
     }
 
     rrdset_set_update_every_s(st, 600);
     min_update_every = __atomic_load_n(&st->rrdhost->stream.rcv.min_update_every, __ATOMIC_RELAXED);
     if(min_update_every != valid_update_every) {
-        fprintf(stderr, "%s: receiver minimum increased from %ld to %u seconds\n",
-                __FUNCTION__, valid_update_every, min_update_every);
+        fprintf(stderr, "%s: receiver minimum increased from %" PRId64 " to %u seconds\n",
+                __FUNCTION__, (int64_t)valid_update_every, min_update_every);
         rc = 1;
     }
